+++ contrib/localnet/orchestrator/smoketest/runner/setup_bitcoin.go
@@ -2,7 +2,6 @@ package runner
 
 import (
 	"encoding/hex"
-	"fmt"
 	"strings"
 	"time"
 
@@ -10,37 +9,62 @@ import (
 	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcd/rpcclient"
 	"github.com/btcsuite/btcutil"
-	"github.com/zeta-chain/zetacore/contrib/localnet/orchestrator/smoketest/utils"
 )
 
-func (sm *SmokeTestRunner) SetupBitcoin() {
-	utils.LoudPrintf("Setup Bitcoin\n")
+func (sm *SmokeTestRunner) SetupBitcoinAccount() {
+	sm.Logger.Print("⚙️ setting up Bitcoin account")
 	startTime := time.Now()
 	defer func() {
-		fmt.Printf("Bitcoin setup took %s\n", time.Since(startTime))
+		sm.Logger.Print("✅ Bitcoin account setup in %s\n", time.Since(startTime))
 	}()
 
 	btc := sm.BtcRPCClient
-	_, err := btc.CreateWallet("smoketest", rpcclient.WithCreateWalletBlank())
+	_, err := btc.CreateWallet(sm.Name, rpcclient.WithCreateWalletBlank())
 	if err != nil {
 		if !strings.Contains(err.Error(), "Database already exists") {
 			panic(err)
 		}
 	}
 
+	sm.setBtcAddress()
+
+	err = btc.ImportAddress(sm.BTCTSSAddress.EncodeAddress())
+	if err != nil {
+		panic(err)
+	}
+
+	// mine some blocks to get some BTC into the deployer address
+	_, err = btc.GenerateToAddress(101, sm.BTCDeployerAddress, nil)
+	if err != nil {
+		panic(err)
+	}
+
+	_, err = btc.GenerateToAddress(4, sm.BTCDeployerAddress, nil)
+	if err != nil {
+		panic(err)
+	}
+}
+
+// setBtcAddress
+func (sm *SmokeTestRunner) setBtcAddress() {
 	skBytes, err := hex.DecodeString(sm.DeployerPrivateKey)
 	if err != nil {
 		panic(err)
 	}
+
+	// TODO: support non regtest chain
+	// https://github.com/zeta-chain/node/issues/1482
 	sk, _ := btcec.PrivKeyFromBytes(btcec.S256(), skBytes)
 	privkeyWIF, err := btcutil.NewWIF(sk, &chaincfg.RegressionNetParams, true)
 	if err != nil {
 		panic(err)
 	}
-	err = btc.ImportPrivKeyRescan(privkeyWIF, "deployer", true)
+
+	err = sm.BtcRPCClient.ImportPrivKeyRescan(privkeyWIF, sm.Name, true)
 	if err != nil {
 		panic(err)
 	}
+
 	sm.BTCDeployerAddress, err = btcutil.NewAddressWitnessPubKeyHash(
 		btcutil.Hash160(privkeyWIF.PrivKey.PubKey().SerializeCompressed()),
 		&chaincfg.RegressionNetParams,
@@ -48,48 +72,6 @@ func (sm *SmokeTestRunner) SetupBitcoin() {
 	if err != nil {
 		panic(err)
 	}
-	fmt.Printf("BTCDeployerAddress: %s\n", sm.BTCDeployerAddress.EncodeAddress())
 
-	err = btc.ImportAddress(sm.BTCTSSAddress.EncodeAddress())
-	if err != nil {
-		panic(err)
-	}
-	_, err = btc.GenerateToAddress(101, sm.BTCDeployerAddress, nil)
-	if err != nil {
-		panic(err)
-	}
-	bal, err := btc.GetBalance("*")
-	if err != nil {
-		panic(err)
-	}
-	_, err = btc.GenerateToAddress(4, sm.BTCDeployerAddress, nil)
-	if err != nil {
-		panic(err)
-	}
-	bal, err = btc.GetBalance("*")
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("balance: %f\n", bal.ToBTC())
-
-	bals, err := btc.GetBalances()
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("balances: \n")
-	fmt.Printf("  mine (Deployer): %+v\n", bals.Mine)
-	if bals.WatchOnly != nil {
-		fmt.Printf("  watchonly (TSSAddress): %+v\n", bals.WatchOnly)
-	}
-	fmt.Printf("  TSS Address: %s\n", sm.BTCTSSAddress.EncodeAddress())
-	go func() {
-		// keep bitcoin chain going
-		for {
-			_, err = btc.GenerateToAddress(4, sm.BTCDeployerAddress, nil)
-			if err != nil {
-				fmt.Println(err)
-			}
-			time.Sleep(5 * time.Second)
-		}
-	}()
+	sm.Logger.Info("BTCDeployerAddress: %s", sm.BTCDeployerAddress.EncodeAddress())
 }