+++ contrib/localnet/orchestrator/smoketest/smoketests/test_erc20_deposit.go
@@ -3,7 +3,6 @@ package smoketests
 import (
 	"fmt"
 	"math/big"
-	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	ethcommon "github.com/ethereum/go-ethereum/common"
@@ -13,55 +12,27 @@ import (
 )
 
 func TestERC20Deposit(sm *runner.SmokeTestRunner) {
-	startTime := time.Now()
-	defer func() {
-		fmt.Printf("test finishes in %s\n", time.Since(startTime))
-	}()
-	utils.LoudPrintf("Deposit USDT ERC20 into ZEVM\n")
+	sm.DepositERC20()
+}
 
+func TestMultipleERC20Deposit(sm *runner.SmokeTestRunner) {
 	initialBal, err := sm.USDTZRC20.BalanceOf(&bind.CallOpts{}, sm.DeployerAddress)
 	if err != nil {
 		panic(err)
 	}
-	txhash := sm.DepositERC20(big.NewInt(1e18), []byte{})
-	utils.WaitCctxMinedByInTxHash(txhash.Hex(), sm.CctxClient)
+	txhash := MultipleDeposits(sm, big.NewInt(1e9), big.NewInt(3))
+	cctxs := utils.WaitCctxsMinedByInTxHash(sm.Ctx, txhash.Hex(), sm.CctxClient, 3, sm.Logger, sm.CctxTimeout)
+	if len(cctxs) != 3 {
+		panic(fmt.Sprintf("cctxs length is not correct: %d", len(cctxs)))
+	}
 
+	// check new balance is increased by 1e9 * 3
 	bal, err := sm.USDTZRC20.BalanceOf(&bind.CallOpts{}, sm.DeployerAddress)
 	if err != nil {
 		panic(err)
 	}
-
-	diff := big.NewInt(0)
-	diff.Sub(bal, initialBal)
-
-	fmt.Printf("balance of deployer on USDT ZRC20: %d\n", bal)
-	supply, err := sm.USDTZRC20.TotalSupply(&bind.CallOpts{})
-	if err != nil {
-		panic(err)
-	}
-	fmt.Printf("supply of USDT ZRC20: %d\n", supply)
-	if diff.Int64() != 1e18 {
-		panic("balance is not correct")
-	}
-
-	utils.LoudPrintf("Same-transaction multiple deposit USDT ERC20 into ZEVM\n")
-	initialBal, err = sm.USDTZRC20.BalanceOf(&bind.CallOpts{}, sm.DeployerAddress)
-	if err != nil {
-		panic(err)
-	}
-	txhash = MultipleDeposits(sm, big.NewInt(1e9), big.NewInt(10))
-	cctxs := utils.WaitCctxsMinedByInTxHash(txhash.Hex(), sm.CctxClient, 10)
-	if len(cctxs) != 10 {
-		panic(fmt.Sprintf("cctxs length is not correct: %d", len(cctxs)))
-	}
-
-	// check new balance is increased by 1e9 * 10
-	bal, err = sm.USDTZRC20.BalanceOf(&bind.CallOpts{}, sm.DeployerAddress)
-	if err != nil {
-		panic(err)
-	}
-	diff = big.NewInt(0).Sub(bal, initialBal)
-	if diff.Int64() != 1e10 {
+	diff := big.NewInt(0).Sub(bal, initialBal)
+	if diff.Int64() != 3e9 {
 		panic(fmt.Sprintf("balance difference is not correct: %d", diff.Int64()))
 	}
 }
@@ -75,47 +46,36 @@ func MultipleDeposits(sm *runner.SmokeTestRunner, amount, count *big.Int) ethcom
 
 	fullAmount := big.NewInt(0).Mul(amount, count)
 
-	// mint
-	tx, err := sm.USDTERC20.Mint(sm.GoerliAuth, fullAmount)
-	if err != nil {
-		panic(err)
-	}
-	receipt := utils.MustWaitForTxReceipt(sm.GoerliClient, tx)
-	if receipt.Status == 0 {
-		panic("mint failed")
-	}
-	fmt.Printf("Mint receipt tx hash: %s\n", tx.Hash().Hex())
-
 	// approve
-	tx, err = sm.USDTERC20.Approve(sm.GoerliAuth, depositorAddr, fullAmount)
+	tx, err := sm.USDTERC20.Approve(sm.GoerliAuth, depositorAddr, fullAmount)
 	if err != nil {
 		panic(err)
 	}
-	receipt = utils.MustWaitForTxReceipt(sm.GoerliClient, tx)
+	receipt := utils.MustWaitForTxReceipt(sm.Ctx, sm.GoerliClient, tx, sm.Logger, sm.ReceiptTimeout)
 	if receipt.Status == 0 {
 		panic("approve failed")
 	}
-	fmt.Printf("USDT Approve receipt tx hash: %s\n", tx.Hash().Hex())
+	sm.Logger.Info("USDT Approve receipt tx hash: %s", tx.Hash().Hex())
 
 	// deposit
 	tx, err = depositor.RunDeposits(sm.GoerliAuth, sm.DeployerAddress.Bytes(), sm.USDTERC20Addr, amount, []byte{}, count)
 	if err != nil {
 		panic(err)
 	}
-	receipt = utils.MustWaitForTxReceipt(sm.GoerliClient, tx)
+	receipt = utils.MustWaitForTxReceipt(sm.Ctx, sm.GoerliClient, tx, sm.Logger, sm.ReceiptTimeout)
 	if receipt.Status == 0 {
 		panic("deposits failed")
 	}
-	fmt.Printf("Deposits receipt tx hash: %s\n", tx.Hash().Hex())
+	sm.Logger.Info("Deposits receipt tx hash: %s", tx.Hash().Hex())
 
 	for _, log := range receipt.Logs {
 		event, err := sm.ERC20Custody.ParseDeposited(*log)
 		if err != nil {
 			continue
 		}
-		fmt.Printf("Multiple deposit event: \n")
-		fmt.Printf("  Amount: %d, \n", event.Amount)
+		sm.Logger.Info("Multiple deposit event: ")
+		sm.Logger.Info("  Amount: %d, ", event.Amount)
 	}
-	fmt.Printf("gas limit %d\n", sm.ZevmAuth.GasLimit)
+	sm.Logger.Info("gas limit %d", sm.ZevmAuth.GasLimit)
 	return tx.Hash()
 }